@@ -11,15 +11,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.16",
-    "@angular/common": "~12.2.16",
-    "@angular/compiler": "~12.2.16",
-    "@angular/core": "~12.2.16",
-    "@angular/forms": "~12.2.16",
-    "@angular/platform-browser": "~12.2.16",
-    "@angular/platform-browser-dynamic": "~12.2.16",
-    "@angular/router": "~12.2.16",
-    "@syncfusion/ej2-angular-navigations": "^20.1.57",
+    "@angular/animations": "~13.3.10",
+    "@angular/common": "~13.3.10",
+    "@angular/compiler": "~13.3.10",
+    "@angular/core": "~13.3.10",
+    "@angular/forms": "~13.3.10",
+    "@angular/platform-browser": "~13.3.10",
+    "@angular/platform-browser-dynamic": "~13.3.10",
+    "@angular/router": "~13.3.10",
+    "@syncfusion/ej2-angular-charts": "^20.1.59",
+    "@syncfusion/ej2-angular-dropdowns": "^20.1.58",
+    "@syncfusion/ej2-angular-grids": "^20.1.58",
+    "@syncfusion/ej2-angular-layouts": "^20.1.58",
+    "@syncfusion/ej2-angular-navigations": "^20.1.58",
+    "@syncfusion/ej2-angular-popups": "^20.1.58",
+    "@syncfusion/ej2-layouts": "^20.1.58",
     "bootstrap": "^5.1.3",
     "jquery": "^3.6.0",
     "ng-cdbangular": "^1.0.0",
@@ -29,9 +35,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.17",
-    "@angular/cli": "~12.2.17",
-    "@angular/compiler-cli": "~12.2.16",
+    "@angular-devkit/build-angular": "~13.3.7",
+    "@angular/cli": "~13.3.7",
+    "@angular/compiler-cli": "~13.3.10",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
@@ -45,6 +51,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.6.4"
   }
 }
