@@ -18,13 +18,14 @@ export class AuthComponent implements OnInit {
   }
   login(userInput: { email: string, password: string }) {
     this.authService.login(userInput.email, userInput.password).then(res => {
-      this.router.navigate(['/viechle']);
+      this.router.navigate(['ws/dashboard']);
+     
     }).catch(e => console.log(e.message));
   }
   signup(userInput: { email: string, password: string })
  {
     this.authService.signup(userInput.email, userInput.password).then(res => {
-      this.router.navigate(['/viechle']);
+      this.router.navigate(['ws/dashboard']);
     }).catch(e => console.log(e.message));
    
   }
