@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormArray, FormControl, FormGroup, RequiredValidator, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { CommandModel, GridComponent, PageSettingsModel } from '@syncfusion/ej2-angular-grids';
 import { Viechle } from 'src/app/model/viechle.interface';
 import { ViechleService } from '../viechle.service';
 
@@ -10,16 +11,16 @@ import { ViechleService } from '../viechle.service';
   styleUrls: ['./add-viechle.component.css']
 })
 export class AddViechleComponent implements OnInit {
+  public data: Viechle[] = [];
+  public pageSettings!: PageSettingsModel;
+  public commands!: CommandModel[];
+  @ViewChild('grid') public grid!: GridComponent;
   form = new FormGroup({
     model: new FormControl('',[Validators.required]),
     type: new FormControl('',[Validators.required]),
     licenceNum: new FormControl('',[Validators.required]),
     viechles: new FormArray([
-      new FormGroup({
-        model: new FormControl({value: 'model123', disabled:true},),
-        type: new FormControl({value: 'type1', disabled:true}),
-        licenceNum: new FormControl({value: 'ldkjfa2847835', disabled:true}),
-      })
+     
     ])
   })
   get viechleArray():FormArray {
@@ -37,6 +38,10 @@ export class AddViechleComponent implements OnInit {
   constructor(private router:Router,private viechleService:ViechleService ) { }
 
   ngOnInit(): void {
+    this.pageSettings = { pageSize: 6 };
+    this.commands = [
+    { buttonOption: { content: 'x', cssClass: '' } }
+    ];
   }
   removeViechle(index:number) {
     this.viechleArray.removeAt(index);
@@ -52,27 +57,42 @@ export class AddViechleComponent implements OnInit {
         type: new FormControl({ value: type, disabled: true }),
         licenceNum: new FormControl({ value: licenceNum, disabled: true }),
       });
-      this.viechleArray.push(newViechleGroup);
+      this.viechleArray.clear();
+      this.viechleArray.push(newViechleGroup)
+      this.data.push(newViechleGroup.value)
 
 
     }
 
   }
   addViechlesToDb() {
-    const viechles = this.viechleArray.value as Viechle[];
+    const viechles = this.data as Viechle[];
+    console.log(this.data);
     this.viechleService.addViechles(viechles).subscribe(errorRes => {
       if (errorRes == '') {
-        this.router.navigate(['/viechle']);
+       
+         this.router.navigate(['/ws/viechle']);
       }
       else {
-        console.log('error is:',errorRes);
+        console.log('error is:', errorRes);
+        
       }
-    
-      
     });
   }
   cancel() {
-    this.router.navigate(['/viechle']);
+    this.router.navigate(['/ws/viechle']);
   }
 
+  commandClick(args: any) {
+    const argsData = args.rowData;
+    const oldData = this.data;
+    this.data = [];
+    oldData.forEach(d => {
+      if (d.licenceNum == argsData.licenceNum && d.model == argsData.model && d.type == argsData.type) {
+      
+      } else {
+      this.data.push(d);  
+  }
+})
+  }
 }
