@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import { Viechle } from 'src/app/model/viechle.interface';
 import { ViechleService } from '../viechle.service';
@@ -10,16 +10,28 @@ import { ViechleService } from '../viechle.service';
 })
 export class ViechleListComponent implements OnInit {
   viechles: Viechle[] = [];
+  public data: Viechle[]=[];
+  columnData: { field: string, headerText: string }[] = [
+    { field: 'model', headerText: 'Model' },
+    { field: 'type', headerText: 'Type' },
+    { field: 'licenceNum', headerText: 'Licence plate' },
+  ];
   constructor(private viechleService:ViechleService , private router:Router) { }
 
   ngOnInit(): void {
     this.viechleService.fetchViechles().subscribe(res => {
       this.viechles = res;
     })
-    
   }
-  viewViechle(id:string) {
-    this.router.navigate(['viechle',id])
+  viewViechle(rowData: any) {
+  
+   this.router.navigate(['/ws/viechle',rowData.id])
   }
-
+  deleteViechle(rowData: any) {
+    if (window.confirm('are you sure')) {
+      this.viechleService.deleteViechle(rowData.id).then(res => {
+        this.ngOnInit();
+      })
+    }
+  }
 }
